+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#new)
     * `FormInterface` now implements `RenderInterface` 
+    * Added new `FormInterface::getTask()` method which reads the task from `form.task` in the blueprint 
 
 # v1.6.5
 ## 04/15/2019